+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 #### New Features
 - Added support for `DataFrame.backfill`, `DataFrame.bfill`, `Series.backfill`, and `Series.bfill`.
 - Added support for `DataFrame.compare` and `Series.compare` with default parameters.
+- Added support for `Series.dt.microsecond` and `Series.dt.nanosecond`.
 
 #### Improvements
 - Removed the public preview warning message upon importing Snowpark pandas.